# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Heat Pump FDD API",
    "description": "Diagnose vapour-compression heat-pump faults from cycle features or simulated conditions.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/evidence/runs": {
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "Paginated results.csv rows",
        "operationId": "getEvidenceRuns",
        "parameters": [
          {
            "name": "experiment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Experiment"
            }
          },
          {
            "name": "protocol",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Protocol"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Label"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceRuns"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/confusion": {
      "get": {
        "tags": [
          "evidence"
        ],
        "summary": "Confusion matrix for a named protocol",
        "operationId": "getEvidenceConfusion",
        "parameters": [
          {
            "name": "protocol",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "holdout-test",
              "title": "Protocol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceConfusion"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EvidenceConfusion": {
        "properties": {
          "protocol": {
            "type": "string",
            "title": "Protocol"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          },
          "matrix": {
            "items": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "array",
            "title": "Matrix"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "protocol",
          "labels",
          "matrix",
          "source"
        ],
        "title": "EvidenceConfusion"
      },
...
      "EvidenceRuns": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/RunRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "csv_path": {
            "type": "string",
            "title": "Csv Path",
            "default": "outputs/results.csv"
          },
          "experiments": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Experiments"
          },
          "protocols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Protocols"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Models"
          },
          "labels": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Labels"
          }
        },
        "type": "object",
        "required": [
          "total",
          "offset",
          "limit",
          "rows"
        ],
        "title": "EvidenceRuns"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunRow": {
        "properties": {
          "experiment": {
            "type": "string",
            "title": "Experiment"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol"
          },
          "reference": {
            "type": "string",
            "title": "Reference"
          },
          "features": {
            "type": "string",
            "title": "Features"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "n": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "experiment",
          "protocol",
          "reference",
          "features",
          "model",
          "label",
          "metric",
          "value"
        ],
        "title": "RunRow"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}